code/requirements.txt
accelerate==0.31.0
annotated-types==0.7.0
appdirs==1.4.4
atomicwrites==1.4.0
av==11.0.0
DataProperty==1.0.1
datasets==2.20.0
decord==0.6.0
deepspeed==0.13.1
docker-pycreds==0.4.0
einops==0.7.0
et-xmlfile==1.1.0
evaluate==0.4.2
ftfy==6.1.3
fvcore==0.1.5.post20221221
graphviz==0.20.3
hiddenlayer==0.3
hjson==3.1.0
huggingface-hub==0.23.4
inflection==0.5.1
iopath==0.1.10
jsonlines==4.0.0
lightning-utilities==0.11.3.post0
lm_eval==0.4.1
mbstrdecoder==1.1.3
mkl-service==2.4.0
multiprocess==0.70.16
ninja==1.11.1.1
onnx==1.16.2
opencv-python==4.8.1.78
pathvalidate==3.2.0
peft==0.11.1
Pillow==10.1.0
ply==3.11
portalocker==2.10.0
protobuf==3.20.3
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1-modules==0.2.8
pybind11==2.13.1
pydantic==2.7.4
pydantic_core==2.18.4
PyDispatcher==2.0.5
pyls-spyder==0.4.0
pynvml==11.5.0
PyQt5==5.15.10
PyQtWebEngine==5.15.6
pytablewriter==1.2.0
rouge_score==0.1.2
sacrebleu==2.4.2
safetensors==0.4.3
sentry-sdk==2.7.0
setproctitle==1.3.3
sqlitedict==2.1.0
tabledata==1.3.3
tcolorpy==0.1.6
tensorboardX==2.6.2.2
termcolor==2.4.0
timm==0.4.12
tokenizers==0.15.2
torch-tb-profiler==0.4.3
torchmetrics==1.4.0.post0
tqdm-multiprocess==0.0.11
transformers==4.36.1
typepy==1.3.2
wandb==0.16.2
wcwidth==0.2.13
webencodings==0.5.1
xxhash==3.4.1
yacs==0.1.8
zstandard==0.22.0
